refactor(build): hoist MemoryQuery from me-query to me-types
MemoryQuery is a pure data + builder DTO (chrono + sibling me-types
vocabulary only, zero me-query-internal dependencies) whose entire
sibling search vocabulary (SearchQuery, SearchResult, QueryResponse,
SearchMode) already lives in me-types::types::search. Leaving it in
me-query forced an illegal L3-to-L3 sibling edge from the forthcoming
me-archive crate (Wave 2 #816 / S4, sub-PR 3a).

Moves the type (+ its unit tests + doctest) into
me-types/src/types/memory_query.rs, re-exported as the single
canonical path me_types::types::search::MemoryQuery. me-query
re-exports it so me_query::MemoryQuery keeps resolving; the facade's
crate::search::MemoryQuery / memory_engine::search::MemoryQuery /
memory_engine::MemoryQuery public paths are unchanged. Drops
me-query's now-empty archive feature (its only gate was
MemoryQuery::include_archives, which moved with the type) and the
matching me-query/archive forward from the facade's archive feature.

Pure relocation: no behavior change, no public API change.
