+++ memory-engine/lib/me-query/Cargo.toml
@@ -24,12 +24,6 @@ tokio = { version = "1", features = ["rt"] }
 # `assemble_results`'s failed-to-load-fact warning (moved verbatim from the facade).
 tracing = "0.1"
 
-[features]
-# `MemoryQuery::include_archives` (moved from the facade's `search/query.rs`) is
-# `#[cfg(feature = "archive")]`-gated; forward so `--all-features` compiles it. No
-# other moved item touches an archive-gated field.
-archive = ["me-types/archive"]
-
 [dev-dependencies]
 # Backend harness for the re-pointed behavior tests (XR-4d) — they exercise the
 # live port_hybrid_search against a real SqliteBackend.
