+++ memory-engine/lib/me-query/src/lib.rs
@@ -4,16 +4,18 @@
 //! Extracted from the facade in Wave 2 #816 / S4, sub-PR 2. The engine's
 //! `MemoryEngine::{query, execute_query}` are thin delegates over the free
 //! functions here.
+//!
+//! `MemoryQuery` moved down to `me-types` in sub-PR 3a (it is a pure data +
+//! builder DTO with zero `me-query`-internal dependencies — its sibling search
+//! vocabulary already lived in `me-types::types::search`, and the L3-to-L3 edge
+//! it forced onto the forthcoming `me-archive` crate was illegal). Re-exported
+//! below so `me_query::MemoryQuery` keeps resolving unchanged for every existing
+//! caller (the facade, this crate's own `execute.rs`, and downstream tests).
 #![cfg_attr(test, allow(clippy::unwrap_used))]
 
 pub mod execute;
 pub mod hybrid;
-pub mod query;
 
-// `execute::query` (the free fn) is deliberately NOT flat-re-exported here: it would
-// collide with the `query` module (holding `MemoryQuery`) at the crate root, exactly
-// the naming tension the facade's original `search::query`/`engine::query` module
-// split existed to avoid. Callers reach it as `me_query::execute::query(...)`.
 pub use execute::{QueryExecution, execute_query};
 pub use hybrid::{port_hybrid_search, rrf_merge};
-pub use query::MemoryQuery;
+pub use me_types::types::search::MemoryQuery;
