+++ memory-engine/lib/me-types/src/types/mod.rs
@@ -15,6 +15,10 @@ mod facts;
 pub mod forgetting;
 /// Inspection statistics + dump-format DTOs.
 pub mod inspect;
+/// [`MemoryQuery`](memory_query::MemoryQuery) — relocated from `me-query` (Wave 2
+/// #816 / S4, sub-PR 3a). Private: re-exported as the single canonical path
+/// `types::search::MemoryQuery` from [`search`] rather than flat/duplicated here.
+mod memory_query;
 mod provenance;
 mod relation;
 mod scope;
