+++ memory-engine/lib/me-types/src/types/search.rs
@@ -1,16 +1,24 @@
 //! Search/query result vocabulary — the DTOs the retrieval surface speaks.
 //!
 //! Pure data types (no `rusqlite`, no search logic): the query input
-//! (`SearchQuery`), the result (`SearchResult`/`MatchType`), the response
-//! (`QueryResponse`/`QueryDiagnostics`), the `SearchMode` selector, and the
-//! `RRF_K` fusion constant. Homed in `me-types` (Wave 2 #816) so the consumer
-//! traits (`Reranker` returns `Vec<SearchResult>`), the backend, and the query
-//! crate all share one definition. The fusion *logic* (`rrf_merge`, hybrid search)
-//! stays in the retrieval layer.
+//! (`SearchQuery`), the fluent query builder (`MemoryQuery`, re-exported from
+//! [`super::memory_query`] — Wave 2 #816 / S4, sub-PR 3a), the result
+//! (`SearchResult`/`MatchType`), the response (`QueryResponse`/`QueryDiagnostics`),
+//! the `SearchMode` selector, and the `RRF_K` fusion constant. Homed in `me-types`
+//! (Wave 2 #816) so the consumer traits (`Reranker` returns `Vec<SearchResult>`),
+//! the backend, and the query crate all share one definition. The fusion *logic*
+//! (`rrf_merge`, hybrid search) stays in the retrieval layer.
 
 use chrono::{DateTime, Utc};
 
 use crate::types::{Fact, FactType};
+
+/// The fluent query-builder DTO. Lives in its own file ([`super::memory_query`])
+/// for size hygiene; re-exported here so it sits alongside its sibling search
+/// vocabulary at the single canonical path `me_types::types::search::MemoryQuery`
+/// (Wave 2 #816 / S4, sub-PR 3a — relocated from `me-query` to kill an illegal
+/// L3-to-L3 sibling edge from the forthcoming `me-archive` crate).
+pub use super::memory_query::MemoryQuery;
 
 /// How to combine search sources.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
